working through loadSamples function.

+++ include/SigMFDataType.h
@@ -51,9 +51,9 @@ public:
     Endianness getEndianness() const;
 
     // helpers for interpreting data on disk.
-    size_t getPrimitiveByteCount() const;     // How many bytes is one primitive.
-    size_t getPrimitivesPerSample() const;    // How many primitives make up a single sample.
-    size_t getBytesPerSample() const;         // How many bytes make up a single sample.
+    int64_t getPrimitiveByteCount() const;     // How many bytes is one primitive.
+    int64_t getPrimitivesPerSample() const;    // How many primitives make up a single sample.
+    int64_t getBytesPerSample() const;         // How many bytes make up a single sample.
 
 // Hidden member variables.
 private:
